SHARD-2420 : Fixes archiver issue where successful receipts get overwritten by newer failure receipts (#242)
* feat: add receipt overwrite protection

Implements a mechanism to prevent overwriting valid receipts with potentially invalid ones, improving data integrity.
Introduces a check to ensure that an existing receipt with a success status (1) cannot be overwritten by a new receipt, even if the new receipt has a failure status (0).
Adds logging for duplicate receipts to aid in debugging and monitoring.

* feat: improve error handling and have it config enabled

* feat: renames config flag for duplicate receipt check

Renames the `enableStatusLogic` configuration flag to
`enableDuplicateReceiptsCheck` to more accurately reflect its
purpose: enabling or disabling the duplicate receipt check in
`storeReceiptData` to prevent overwriting successful receipts
with failures.

* feat: improves data logging and duplicate receipt handling.

Enhances the reliability of data logging by adding a try-catch block to the `writeToLog` function to handle potential errors during the writing process.

Provides additional logging for duplicate receipts, including the content of both the existing and incoming receipts, to aid in debugging and analysis. This improves visibility into potential data inconsistencies and helps to identify the root cause of duplicate receipts.